adds margin to the bottom of the elements before footer
due to the technical constraints of using only css,
and the fact there is no css selector for previous sibling,
the approach taken was to hardcode the css selectors while
overqualifying them. still is not necessarily safe that there are no side-effects
but it doesn't seem to imply any currently.

closes #1